requirements.txt
chardet~=5.2.0
requests~=2.31.0
beautifulsoup4~=4.12.3
openai~=1.14.1
python-dotenv~=1.0.1